Updated UC9

+++ src/main/java/domain/entity/UserRating.java
@@ -5,17 +5,19 @@ import java.time.Instant;
 /**
  * Domain entity: rating that a user gives to a recipe.
  * UC9 Favourite / Rate Recipe.
+ *
+ * stars: 0.0 .. 5.0 in 0.5 increments.
  */
 public class UserRating {
 
     private Long id;
     private Long userId;
     private Long recipeId;
-    private int stars;          // integer 1..5
+    private double stars;          // 0.0 .. 5.0, step 0.5
     private Instant updatedAt;
 
     public UserRating(Long id, Long userId, Long recipeId,
-                      int stars, Instant updatedAt) {
+                      double stars, Instant updatedAt) {
         this.id = id;
         this.userId = userId;
         this.recipeId = recipeId;
@@ -33,8 +35,8 @@ public class UserRating {
     public Long getRecipeId() { return recipeId; }
     public void setRecipeId(Long recipeId) { this.recipeId = recipeId; }
 
-    public int getStars() { return stars; }
-    public void setStars(int stars) { this.stars = stars; }
+    public double getStars() { return stars; }
+    public void setStars(double stars) { this.stars = stars; }
 
     public Instant getUpdatedAt() { return updatedAt; }
     public void setUpdatedAt(Instant updatedAt) { this.updatedAt = updatedAt; }
